+++ generators/client-2/templates/gulp/copy.js
@@ -12,17 +12,30 @@ var handleErrors = require('./gulp/handleErrors');
 var config = require('./config');
 
 module.exports = {<% if(enableTranslation) { /* copy i18n folders only if translation is enabled */ %>
-    i18n: i18n,<% } %>
+    i18n: i18n,
+    languages: languages,<% } %>
     fonts: fonts,
     common: common,
     deps: deps
 }
 <% if(enableTranslation) { %>
+var yorc = require('../.yo-rc.json')['generator-jhipster'];
+
 function i18n() {
     return gulp.src(config.app + 'i18n/**')
         .pipe(plumber({errorHandler: handleErrors}))
         .pipe(changed(config.dist + 'i18n/'))
         .pipe(gulp.dest(config.dist + 'i18n/'));
+}
+
+function languages() {
+    var locales = yorc.languages.map(function (locale) {
+        return config.bower + 'angular-i18n/angular-locale_' + locale + '.js';
+    });
+    return gulp.src(locales)
+        .pipe(plumber({errorHandler: handleErrors}))
+        .pipe(changed(config.app + 'i18n/'))
+        .pipe(gulp.dest(config.app + 'i18n/'));
 }
 <% } %>
 function fonts() {
